Rename kas::class -> classes
