+++ crates/kas-core/src/classes.rs
